added base, gps and teleop drone interfaces
