+++ examples/example1.py
@@ -12,13 +12,13 @@ drone_interface.takeoff(3, 2)
 print("Takeoff completed\n")
 sleep(1)
 
-drone_interface.go_to(0, 5, 3)
-drone_interface.go_to(5, 0, 2)
+drone_interface.goto(0, 5, 3, speed=1.0)
+drone_interface.goto(5, 0, 2, speed=1.0)
 
 drone_interface.follow_path([[5, 0, 3],
-                                [5, 5, 3],
-                                [0, 5, 3],
-                                [0, 0, 3]], 5)
+                             [5, 5, 3],
+                             [0, 5, 3],
+                             [0, 0, 3]], 5)
 
 print("Path finished")
