+++ python_interface/behaviour_actions/land_behaviour.py
@@ -43,13 +43,13 @@ from as2_msgs.action import Land
 from ..behaviour_actions.action_handler import ActionHandler
 
 if typing.TYPE_CHECKING:
-    from ..drone_interface import DroneInterface
+    from ..drone_interface_base import DroneInterfaceBase
 
 
 class SendLand(ActionHandler):
     """Land action"""
 
-    def __init__(self, drone: 'DroneInterface', speed: float = 0.0, wait_result: bool = True) -> None:
+    def __init__(self, drone: 'DroneInterfaceBase', speed: float = 0.0, wait_result: bool = True) -> None:
         self._action_client = ActionClient(drone, Land, 'LandBehaviour')
 
         goal_msg = Land.Goal()
